# client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: fbstachura@gmail.com
  description: StampWallet API Server REST Specification
  title: StampWallet API Server
  version: 0.1.0
servers:
- url: /
paths:
...
  /user/cards:
    get:
      description: This endpoint is used to retrieve a list of currently logged in
        user's cards.
      operationId: getUserCards
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserCardsResponse'
          description: Successful operation
      security:
      - sessionToken: []
      summary: Get list of user's cards
      tags:
      - cards
...
  /user/cards/virtual/{businessId}/itemsDefinitions/{itemDefinitionId}:
    post:
      description: This endpoint is used to buy an item for points from the virtual
        card.
      operationId: buyItem
      parameters:
      - description: Public ID of the business which card was requested
        explode: false
        in: path
        name: businessId
        required: true
        schema:
          type: string
        style: simple
      - description: Public ID of the item definition requested by the user
        explode: false
        in: path
        name: itemDefinitionId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostUserVirtualCardItemResponse'
          description: Successful operation
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: No such item or item does not belong to this business
        "401":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: "Item withdrawn, not available etc"
      security:
      - sessionToken: []
      summary: Buy an item
      tags:
      - virtualCards
  /user/cards/virtual/{businessId}/items/{itemId}:
    delete:
      description: "This endpoint is used to return an item, and get back points that\
        \ were spent on that item."
      operationId: deleteItem
      parameters:
      - description: Public ID of the business which card was requested
        explode: false
        in: path
        name: businessId
        required: true
        schema:
          type: string
        style: simple
      - description: Public ID of the item requested to be deleted
        explode: false
        in: path
        name: itemId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                status: OK
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Successful operation
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: No such item
        "401":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Item cannot be returned
      security:
      - sessionToken: []
      summary: Delete an item
      tags:
      - virtualCards
...
  /user/businesses:
    get:
      description: This endpoint is used to search businesses that match the provided
        text query or are close to a specified point.
      operationId: searchBusinesses
      parameters:
      - description: Filter by business name
        explode: true
        in: query
        name: text
        required: false
        schema:
          type: string
        style: form
      - description: Filter by business location
        explode: true
        in: query
        name: location
        required: false
        schema:
          format: ISO 6709
          type: string
        style: form
      - description: Filter by distance from location in meters
        explode: true
        in: query
        name: proximity
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserBusinessesSearchResponse'
          description: Successful operation
        "400":
          content:
            application/json:
              example:
                status: INVALID_REQUEST
              schema:
                $ref: '#/components/schemas/DefaultResponse'
          description: Invalid request
      security:
      - sessionToken: []
      summary: Search businesses
      tags:
      - user
...
components:
  schemas:
    DefaultResponse:
      example:
        message: message
        status: null
      properties:
        status:
          $ref: '#/components/schemas/DefaultResponseStatusEnum'
        message:
          type: string
      type: object
    DefaultResponseStatusEnum:
      enum:
      - OK
      - CREATED
      - NOT_FOUND
      - FORBIDDEN
      - ALREADY_EXISTS
      - UNAUTHORIZED
      - INVALID_REQUEST
      - UNKNOWN_ERROR
      - CONFLICT
      type: string
...
    GetUserCardsResponse:
      example:
        localCards:
        - publicId: d29oU29vdm9pNW9vCg
          name: test card
          type: aWVuZWVnZWVZYWkyCg
          code: "1234567890"
        - publicId: d29oU29vdm9pNW9vCg
          name: test card
          type: aWVuZWVnZWVZYWkyCg
          code: "1234567890"
        virtualCards:
        - businessDetails:
            publicId: cG9leWlWb283c2h1Cg
            name: Business
            description: Coffee shop or whatever
            gpsCoordinates: +48.8577+002.295/
            bannerImageId: b29naG9oVGhhaG42Cg
            iconImageId: b29naG9oVGhhaG42Cg
          points: 0
        - businessDetails:
            publicId: cG9leWlWb283c2h1Cg
            name: Business
            description: Coffee shop or whatever
            gpsCoordinates: +48.8577+002.295/
            bannerImageId: b29naG9oVGhhaG42Cg
            iconImageId: b29naG9oVGhhaG42Cg
          points: 0
      properties:
        localCards:
          items:
            $ref: '#/components/schemas/LocalCardAPIModel'
          type: array
        virtualCards:
          items:
            $ref: '#/components/schemas/ShortVirtualCardAPIModel'
          type: array
      type: object
...
    GetUserBusinessesSearchResponse:
      example:
        businesses:
        - publicId: cG9leWlWb283c2h1Cg
          name: Business
          description: Coffee shop or whatever
          gpsCoordinates: +48.8577+002.295/
          bannerImageId: b29naG9oVGhhaG42Cg
          iconImageId: b29naG9oVGhhaG42Cg
      properties:
        businesses:
          items:
            $ref: '#/components/schemas/ShortBusinessDetailsAPIModel'
          type: array
      type: object
...
    PostUserVirtualCardItemResponse:
      example:
        itemId: aWVSOGplZXdlaW1pCg
      properties:
        itemId:
          type: string
      type: object
...
    ShortBusinessDetailsAPIModel:
      example:
        publicId: cG9leWlWb283c2h1Cg
        name: Business
        description: Coffee shop or whatever
        gpsCoordinates: +48.8577+002.295/
        bannerImageId: b29naG9oVGhhaG42Cg
        iconImageId: b29naG9oVGhhaG42Cg
      properties:
        publicId:
          type: string
        name:
          type: string
        description:
          type: string
        gpsCoordinates:
          type: string
        bannerImageId:
          type: string
        iconImageId:
          type: string
      type: object
...
    LocalCardAPIModel:
      example:
        publicId: d29oU29vdm9pNW9vCg
        name: test card
        type: aWVuZWVnZWVZYWkyCg
        code: "1234567890"
      properties:
        publicId:
          type: string
        name:
          type: string
        type:
          type: string
        code:
          type: string
      type: object
    ShortVirtualCardAPIModel:
      example:
        businessDetails:
          publicId: cG9leWlWb283c2h1Cg
          name: Business
          description: Coffee shop or whatever
          gpsCoordinates: +48.8577+002.295/
          bannerImageId: b29naG9oVGhhaG42Cg
          iconImageId: b29naG9oVGhhaG42Cg
        points: 0
      properties:
        businessDetails:
          $ref: '#/components/schemas/ShortBusinessDetailsAPIModel'
        points:
          type: integer
      type: object
...
  securitySchemes:
    sessionToken:
      bearerFormat: "Just the token string received from the API. In reality consists\
        \ of the token id and the actual token, separated by a colon."
      scheme: bearer
      type: http